# [BPK-2291] Add theming support to BpkText