--- a/UNRELEASED.md
+++ b/UNRELEASED.md
@@ -1,1 +1,6 @@
 # Unreleased
+
+**Added:**
+
+- react-native-bpk-component-text:
+  - Added support for changing the font using a theme prop.