--- a/packages/react-native-bpk-component-text/package.json
+++ b/packages/react-native-bpk-component-text/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "bpk-tokens": "^27.4.7",
-    "prop-types": "^15.5.8"
+    "prop-types": "^15.5.8",
+    "react-native-bpk-theming": "^1.0.91"
   }
 }