@shaundon 2019-03-06
> Do we know if this has any impact on performance?

I could do some measuring, but is there much we can do? This is the standard way we do theming, if we go down a different route here wouldn't it imply a problem with our overall approach to theming?

We could do something like have a new `BpkThemedText` component but then that means we wouldn't be able to target all existing instances of BpkText.