@matthewdavidson 2019-03-06
A quick comparison of 1000 `BpkText` vs 1000 `withTheme(BpkText)` would quell this. The guide on measuring performance on official docs should help: https://facebook.github.io/react-native/docs/performance#profiling

Comparing before and after with regards to memory usage and frame rate should be enough.