@shaundon 2019-03-06
@k0nserv @matthewdavidson just did a quick perf test. I made the following two stories:

```javascript
  .add('Perf test', () => (
    <ScrollView>
      {Array(1000)
        .fill()
        .map((k, i) => (
          <BpkText key={i.toString()}>Developers</BpkText>
        ))}
    </ScrollView>
  ))

  .add('Perf test (themed)', () => (
    <BpkThemeProvider theme={themeAttributes}>
      <ScrollView>
        {Array(1000)
          .fill()
          .map((k, i) => (
            <BpkText key={i.toString()}>Developers</BpkText>
          ))}
      </ScrollView>
    </BpkThemeProvider>
  ))
```

Then I loaded each story with the RN profiler open and compared memory usage (iOS), stutters (Android), UI and JS frames per second while rapidly scrolling through the list.

**iOS**

Without theming:
	60fps for both UI and JS
	RAM constant 263-264mb

With theming:
	60fps for both UI and JS
	RAM constant at 265-6mb


**Android**

Without theming:
	~54fps for both UI and JS
	0 stutters
	

With theming:
	~54fps for both UI and JS
	0 stutters

So as far as I can see, performance is identical. I can dive in and use Android and iOS's native profilers if you think it's needed but this felt sufficient.